+++ mapcraft/src/map/elements/Terrain.java
@@ -18,7 +18,7 @@ package net.sourceforge.mapcraft.map.elements;
  * @version $Revision$
  */
 public class Terrain implements Cloneable {
-    private short   id;
+    private int     id;
     private String  name;
     private String  description;
     private String  imagePath;
@@ -27,14 +27,14 @@ public class Terrain implements Cloneable {
     private int     quality;
 
     public
-    Terrain(short id, String name, String description, String imagePath) {
+    Terrain(int id, String name, String description, String imagePath) {
         this.id = id;
         this.name = name;
         this.description = description;
         this.imagePath = imagePath;
     }
 
-    public short getId() { return id; }
+    public int getId() { return id; }
     public String getName() { return name; }
     public String getDescription() { return description; }
     public String getImagePath() { return imagePath; }
